god help us

+++ D.MG/Models/SettlementType.cs
@@ -4,9 +4,9 @@ namespace Models
 {
     public class SettlementType : BaseEntity
     {
-        public double Downpayment { get; set; }
+        public decimal Downpayment { get; set; }
         public int Installments { get; set; }
-        public double Interest { get; set; }
+        public decimal Interest { get; set; }
 
         public virtual ICollection<Settlement> Settlements { get; set; }
